Restructured navigation component
