+++ app/pages/trabajo.js
@@ -1,6 +1,6 @@
 import Head from 'next/head'
 import Layout from '../components/layout'
-import Nav from '../components/nav'
+import Nav from '../components/navigation/nav'
 import Content from '../components/grid/container'
 
 export default function Home() {
